Get successful build on MacOS
Changed compiler flags that mask 1000s of warnings. Extend int's to intptr_t for pointer aliasing.

+++ vm/asm/location.hpp
@@ -161,7 +161,7 @@ const Location topOfFloatStack	= Location::specialLocation(5);	// only used if U
 class IntFreeList : public PrintableResourceObj {
  protected:
   int			_first;	// the first available integer
-  GrowableArray<int>*	_list;	// the list
+  GrowableArray<intptr_t>*	_list;	// the list
 
   void grow();
