+++ vm/code/nmethod.cpp
@@ -388,8 +388,8 @@ void nmethod::makeZombie(bool clearInlineCaches) {
     std->print_cr("%s nmethod 0x%x becomes zombie", (is_method() ? "normal" : "block"), this);
     if (WizardMode) {
       std->print_cr("entry code sequence:");
-      char* beg = (char*)min(int(specialHandlerCall()), int(entryPoint()), int(verifiedEntryPoint()));
-      char* end = (char*)max(int(specialHandlerCall()), int(entryPoint()), int(verifiedEntryPoint()));
+      char* beg = (char*)min(intptr_t(specialHandlerCall()), intptr_t(entryPoint()), intptr_t(verifiedEntryPoint()));
+      char* end = (char*)max(intptr_t(specialHandlerCall()), intptr_t(entryPoint()), intptr_t(verifiedEntryPoint()));
       Disassembler::decode(beg, end + 10);
     }
   }
